Python/codigo/airtable_api/requirements.txt
pyairtable==1.5.0
python-dotenv==1.0.1 
